requirements.txt
numpy==1.11.1
Django==1.9.8
Pillow==3.3.0
h5py==2.6.0
nltk==3.2.1
scipy
channels==0.17.2
asgiref==1.0.0
asgi-redis==0.14.1
pika==0.10.0
pyyaml==3.12
uwsgi==2.0.13.1
websocket-client==0.37.0
daphne==0.15.0
requests==2.11.1
mysqlclient==1.3.13
matplotlib
ipython
pandas
jupyter
nose
pysqlite